[daisy5] Fix augmentation of form module interfaces
