+++ packages/daisyui5-theme/src/lib/widgets/cally-calendar/index.ts
@@ -1,1 +1,2 @@
+export type { CalendarProps } from './component.svelte'
 export { default as CallyCalendar } from './wrapper.svelte'
