+++ packages/daisyui5-theme/src/lib/widgets/toggle.svelte
@@ -1,18 +1,3 @@
-<script lang="ts" module>
-	import type { HTMLInputAttributes } from 'svelte/elements';
-	import type { WidgetCommonProps } from '@sjsf/form';
-
-	declare module '@sjsf/form' {
-		interface WidgetsAndProps<V> {
-			toggle: WidgetCommonProps<V, HTMLInputAttributes>;
-		}
-
-		interface WidgetValue {
-			toggle: boolean;
-		}
-	}
-</script>
-
 <script lang="ts">
 	import type { WidgetProps } from '@sjsf/form';
